The "dist", "vendor" and "directory" parameters are now optional
This required introducing a "basedir" setting and integrating
the chroot identifier in the underlying "data" dictionnary.

Update the documentation accordingly and use this opportunity
to improve it too.

Misc changes made while implementing the above:
- [bugfix] do not hide failures of debootstrap (with the success of the chroot
  removal)
- [style] use data.foo instead of data['foo'] to improve readability

NOTE: The interface of the macros exported via debootstrap/states.jinja
changed but there are no users at this point, so don't worry about
backwards compat.